Interleave complexity factors
This leads to more even runtimes for shards. Before, if shard size was 5, the first shard would get the 5 fastest tests and the seconds shard would get the 5 slowest tests. Now the first shard will get the 3 fastest tests and the 2 slowest, while the second shard will get the 5 tests in the middle.